# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API de Gestão Escolar",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/professionals/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Filtro de busca",
            "in": "query",
            "required": false,
            "type": "object",
            "schema": {
              "$ref": "#/components/schemas/professional"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/professional"
              }
            }
          }
        }
      }
    },
    "/professionals/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "speciality": {
                    "example": "any"
                  },
                  "contact": {
                    "example": "any"
                  },
                  "phoneNumber": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/students/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Filtro de busca",
            "in": "query",
            "required": false,
            "type": "object",
            "schema": {
              "$ref": "#/components/schemas/student"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/student"
              }
            }
          }
        }
      }
    },
    "/students/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "age": {
                    "example": "any"
                  },
                  "parents": {
                    "example": "any"
                  },
                  "phoneNumber": {
                    "example": "any"
                  },
                  "specialNeeds": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "student": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Bingo Heeler"
          },
          "age": {
            "type": "number",
            "example": 6
          },
          "parents": {
            "type": "string",
            "example": "Bandit Heeler e Chilli Heeler"
          },
          "phoneNumber": {
            "type": "string",
            "example": "48 9696 5858"
          },
          "specialNeeds": {
            "type": "string",
            "example": "Síndrome de Down"
          },
          "status": {
            "type": "string",
            "example": "on"
          }
        },
        "required": [
          "name",
          "age",
          "parents",
          "phoneNumber",
          "specialNeeds"
        ]
      },
      "professional": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Winton Blake"
          },
          "specialty": {
            "type": "string",
            "example": "Fisioterapeuta"
          },
          "contact": {
            "type": "string",
            "example": "w.blake@gmail"
          },
          "phoneNumber": {
            "type": "string",
            "example": "48 9696 5858"
          },
          "status": {
            "type": "string",
            "example": "on"
          }
        },
        "required": [
          "name",
          "specialty",
          "contact",
          "phoneNumber"
        ]
      },
...
    }
  }
}